fix(rta): make every demo wait report what it saw, and gate the directory
The diagnostics rule this branch introduced ("a timeout must report the state it
SAW, so throw via `diagnosedError`") had two live exceptions inside the very
runner the branch instrumented with `setFailureContext`:

  - `demos/run.mjs`'s `waitMediaPlaying` — a byte-identical copy of the one in
    `specs/deeplink.spec.js`, which WAS converted. So a demo take that failed at
    playback left no record at all, which is the failure that silently ruins
    footage that looks fine until playback.
  - `demos/takes/server-switch.js`'s 15s poll for the change-server dialog — a
    hand-rolled loop, so "dialog never appeared" could not be attributed.

Both are fixed at the shape level rather than one-off:

  - `waitMediaPlaying` / `stopPlayback` / `PLAYING_STATES` move into
    `lib/steps.js` and are shared. The two copies were identical apart from the
    message label, which is now a parameter (`deep-link play`, `demo`) — the two
    callers are labelled differently by the harness, since Vitest names the
    spec's test and nothing names a take.
  - the take's poll becomes `ctx.waitFor`, which the runner's toolkit already
    exposed and which already throws through `diagnosedError`.

The ESLint gate widens from the two lib files to `tests/rta/demos/**`. That is
on evidence, not symmetry: while demos sat outside the glob it accumulated the
two waits above, both shipped in the same change set that wrote the rule down,
and it is the directory that grows by adding choreography — which is where new
waits come from. Cost is six one-line disables on genuine fail-fasts (an unknown
server name, the non-demo-host privacy refusal, `land()`'s unsupported screen,
two empty REST lookups, and the stash-cleared outcome check), each carrying its
reason. `specs/` stays out: most spec-level throws are assertions, not timeouts.

Verified: `npm run lint:js` clean (the gate passes with the disables, and fails
without them — confirmed by widening the glob before adding them), `lint:docs`
92 files clean, prettier clean. The hoisted helper is exercised on hardware by
`deeplink.spec.js` under `npm run test:rta`; the demos path needs an operator at
its stop gate and is covered separately.
