+++ docs/architecture/build-and-tooling.md
@@ -440,7 +440,7 @@ Rule for new scripts: net-new top-level CLI scripts go ESM `.js`; plugins and sh
 ### Tooling
 
 - **[`eslint.config.js`](../../eslint.config.js)** — flat config, `@eslint/js` recommended + `eslint-plugin-n` `mixed-esm-and-cjs` preset. Disables `n/no-unpublished-*` (false positive — JellyRock isn't an npm package). Enforces `n/hashbang` for CLI shebangs, `n/prefer-node-protocol` for ESM imports of built-ins, plus standard `no-unused-vars` / `no-var` / `prefer-const`. `eslint-config-prettier` runs LAST to disable formatting rules that would fight Prettier.
-  - It also carries **path-scoped project invariants**, not just generic hygiene — a `files:` block plus `no-restricted-syntax` turns a convention that was previously prose-only into a gate. Current member: RTA waits (`tests/rta/lib/nav.js`, `tests/rta/lib/steps.js`) must throw through `diagnosedError` rather than a bare `new Error`, so a timeout reports the device state it saw ([rta-tests.md](../dev/rta-tests.md#when-a-wait-times-out-it-reports-what-it-saw)). This is the cheap shape for a single-file-scope invariant: no new script, no `package.json` entry, and no CI wiring — it rides `lint:js`, which already has a CI home whose path filter matches, so [`ci-parity-check`](../../scripts/lint/ci-parity-check.js) stays satisfied. Prefer it over a bespoke `scripts/lint/*` checker unless the rule needs to reason across files. A legitimate exception disables the rule on its line **with a reason**.
+  - It also carries **path-scoped project invariants**, not just generic hygiene — a `files:` block plus `no-restricted-syntax` turns a convention that was previously prose-only into a gate. Current member: RTA waits (`tests/rta/lib/nav.js`, `tests/rta/lib/steps.js`, `tests/rta/demos/**`) must throw through `diagnosedError` rather than a bare `new Error`, so a timeout reports the device state it saw ([rta-tests.md](../dev/rta-tests.md#when-a-wait-times-out-it-reports-what-it-saw)). This is the cheap shape for a single-file-scope invariant: no new script, no `package.json` entry, and no CI wiring — it rides `lint:js`, which already has a CI home whose path filter matches, so [`ci-parity-check`](../../scripts/lint/ci-parity-check.js) stays satisfied. Prefer it over a bespoke `scripts/lint/*` checker unless the rule needs to reason across files. A legitimate exception disables the rule on its line **with a reason**.
 - **[`.prettierrc.json`](../../.prettierrc.json)** — anchored to bsfmt style: 2-space indent, single quotes, semis, trailing commas (all), 100-col print width. JSONC parser for `.vscode/*.json` (they have comments).
 - **[`.prettierignore`](../../.prettierignore)** — excludes `package-lock.json` (npm-owned), `locale/` (translation tooling owns), `tasks/jellyfin-server-openapi/` (vendor), `CHANGELOG.md` (CI-controlled), and the formats other tools own (`.bs`/`.brs`/`.xml`/`.md`).
 - **[`vitest.config.js`](../../vitest.config.js)** — picks up `tests/scripts/**/*.test.js` plus `tests/rta/**/*.test.js` (the RTA harness's hardware-free parts, e.g. the registry restore planner). The `.test.js` / `.spec.js` split is load-bearing: `tests/rta/specs/**/*.spec.js` drives a real device and belongs to `vitest.rta.config.js`, so it must never be picked up here. Test files are ESM regardless of source module system (Vitest handles cross-module-system imports).
