+++ tests/rta/CLAUDE.md
@@ -53,7 +53,7 @@ Canonical examples in [`lib/nav.js`](lib/nav.js): `waitOsdUp` (input), `findHome
 
 - `config.js` — `RTA_CONFIG` (demo server, hero movie, seek position, locales). Shared with the store screenshot generator.
 - `screens.js` — the **screen registry**, the single source of truth for both the tests and the screenshots. Add a screen here.
-- `lib/` — `driver` (env + deploy + relaunch), `steps` (press/getVal/waitFor/waitFocused), `seed` (registry seeds), `registry` (whole-registry snapshot/verified restore), `jellyfin` (demo REST), `nav` (per-screen navigation), `diagnostics` (failure-time device-state dump + the `FAILURE_KINDS` registry). The run record it writes into — directories, lifecycle, ledger, summary — lives in [`scripts/run-record.js`](../../scripts/run-record.js), which is shared with the Rooibos runner and knows nothing about devices.
+- `lib/` — `driver` (env + deploy + relaunch), `steps` (press/getVal/waitFor/waitFocused, plus the shared media-player waits used by both the deep-link spec and the demo runner), `seed` (registry seeds), `registry` (whole-registry snapshot/verified restore), `jellyfin` (demo REST), `nav` (per-screen navigation), `diagnostics` (failure-time device-state dump + the `FAILURE_KINDS` registry). The run record it writes into — directories, lifecycle, ledger, summary — lives in [`scripts/run-record.js`](../../scripts/run-record.js), which is shared with the Rooibos runner and knows nothing about devices.
 - `specs/` — the Vitest specs (a `for` loop over `SCREENS`, not `it.each`: `it.each` passes only the case object, so a case can't skip itself at runtime, and content-dependent screens need exactly that).
 - `capture.js`, `setup/` — the `RTA_CAPTURE` raw-capture helper and Vitest global/per-worker setup.
 - `demos/` — hands-free **video-capture** takes (`npm run demo`). `run.mjs` owns the privacy-safe lifecycle (snapshot → record gates → restore + relaunch); each `takes/*.js` declares only its choreography. NOT tests — these drive the device for marketing/PR demos against the public demo server only (the runner refuses any non-demo host).
@@ -66,7 +66,8 @@ Canonical examples in [`lib/nav.js`](lib/nav.js): `waitOsdUp` (input), `findHome
 - **`waitFor`/`waitFocused` throw on timeout — that IS the assertion.** Don't wrap them in `expect`. Use `expect` only for value checks (label text, focus subtype).
 - **Preconditions before actions and reads** — see [the north star](#-the-north-star-establish-the-state-that-makes-an-action-or-a-read-meaningful-before-doing-it) at the top of this file. It is the rule most worth internalizing before you touch a nav.
 - **A timeout must report what it SAW, so throw via `diagnosedError`, never a bare `new Error`.** `lib/diagnostics.js` attaches the state the device was actually in (active view + `loadState`, the app shell's `isLoading` / `isRemoteDisabled`, focused node, row counts, seeded server/user identity) and appends a record to the run's `failures.jsonl` that the run's `close()` folds into `run-meta.json` (see [`scripts/run-record.js`](../../scripts/run-record.js)). The capture runs only at the throw site, after a poll loop has given up — never inside a tick — so it costs nothing on the success path (measured on `.177`: median 21 ms, n=20). This applies to **timeouts**; a fail-fast that already names its cause (the ambiguous-library refusal in `nav.js`) can stay a plain throw.
-  - **This is enforced, not just written down.** An ESLint `no-restricted-syntax` rule scoped to `lib/nav.js` + `lib/steps.js` fails `lint:js` on a bare `throw new Error` there. A legitimate fail-fast disables it on that line **with a reason**. A new lib file that grows a wait should be added to that glob in [`eslint.config.js`](../../eslint.config.js).
+  - **This is enforced, not just written down.** An ESLint `no-restricted-syntax` rule scoped to `lib/nav.js`, `lib/steps.js` and **all of `demos/`** fails `lint:js` on a bare `throw new Error` there. A legitimate fail-fast disables it on that line **with a reason**. A new lib file that grows a wait should be added to that glob in [`eslint.config.js`](../../eslint.config.js); `specs/` stays out, because most spec-level throws are assertions rather than timeouts.
+  - **In a take, reach for `ctx.waitFor` rather than a hand-rolled poll** — it already throws through `diagnosedError`, so you inherit the dump instead of re-deriving it. A take that polls its own loop is the shape that produced both of the unconverted waits this rule was written for.
   - **`kind` comes from the frozen `FAILURE_KINDS` set, never an inline string** — it is the key a flake baseline aggregates by, and an invented slug forks a bucket. An unregistered one is recorded and called out in the run summary rather than silently counted.
   - **Prefer passing state the loop *already read* as `observed`** over re-reading it; that is what turns `2 row(s) present` into `rowTypes=[Chapter, Person]`, i.e. the difference between *late* and *absent*.
   - **`loadState=—` on a detail screen is correct** — that field is `BaseGridView`'s, and `ItemDetails` extends `JRScreen`. Read `detail=<n>` and the shell fields there instead; `input=BLOCKED` means `JRScene.onKeyEvent` was swallowing every key we sent, which is the north-star failure mode made visible.
