+++ tests/rta/specs/deeplink.spec.js
@@ -16,11 +16,19 @@ import { RTA_CONFIG } from '../config.js';
 import { authenticate, getHero, getLibraries, libraryIdFor, firstItemId } from '../lib/jellyfin.js';
 import { seedHome, assertSeedTookEffect } from '../lib/seed.js';
 import { hardRelaunch, ecp } from '../lib/driver.js';
-import { press, waitFor, getActiveVal, waitHome, sleep, hasChildren } from '../lib/steps.js';
-import { diagnosedError, FAILURE_KINDS } from '../lib/diagnostics.js';
+import {
+  press,
+  waitFor,
+  getActiveVal,
+  waitHome,
+  sleep,
+  hasChildren,
+  waitMediaPlaying,
+  stopPlayback,
+  PLAYING_STATES,
+} from '../lib/steps.js';
 
 const LOCALE = RTA_CONFIG.languages[0];
-const PLAYING_STATES = ['startup', 'buffer', 'play', 'pause']; // Roku media-player active states
 // Upper bound on the validate-before-navigate metadata fetch (+ toast). The invalid-id test
 // asserts a NON-event (no navigation), so it must out-wait the fetch; this is a generous ceiling
 // on a LAN fetch to the demo, not a tuned value. If the production fetch ever gains an explicit
@@ -51,47 +59,6 @@ async function castFromHome(contentId) {
   await ecp.sendInput({ params: { contentId } }); // querystring.build URL-encodes | and =
 }
 
-/** Poll the device media-player until it reaches an active playback state. */
-async function waitMediaPlaying(label, timeout = 30000) {
-  const start = Date.now();
-  let last;
-  while (Date.now() - start < timeout) {
-    const mp = await ecp.getMediaPlayer().catch(() => null);
-    last = mp?.state;
-    if (mp && !mp.error && PLAYING_STATES.includes(mp.state)) return;
-    await sleep(1000);
-  }
-  // A timeout, so it reports what it SAW — same rule as the lib waits. "Never
-  // started" alone cannot distinguish a stream that failed to open from a cast the
-  // app never routed, and the shell fields answer that: `input=BLOCKED` or a live
-  // spinner means the app was busy, not that playback died.
-  throw await diagnosedError(
-    `deep-link ${label}: media player never started (last state=${last})`,
-    {
-      kind: FAILURE_KINDS.MEDIA_PLAYER_NOT_STARTED,
-      label: `media player (${label})`,
-      waitedMs: Date.now() - start,
-      observed: { lastPlayerState: last, expected: PLAYING_STATES },
-    },
-  );
-}
-
-/**
- * Stop playback so it can't leak into the next spec. Back on the player stops it
- * (AudioPlayerView/PlayerHostView onKeyEvent "back" → control "stop") — necessary
- * because AUDIO keeps playing while you navigate away (correct music-app UX), so a
- * relaunch-to-Home alone won't silence it. Verify via media-player; retry Back.
- */
-async function stopPlayback() {
-  const start = Date.now();
-  while (Date.now() - start < 10000) {
-    const mp = await ecp.getMediaPlayer().catch(() => null);
-    if (!mp || !PLAYING_STATES.includes(mp.state)) return;
-    await press(ecp.Key.Back);
-    await sleep(1200);
-  }
-}
-
 // An invalid id must never disturb the session: validate-before-navigate toasts and
 // stays put (ADR 0018). Assert we never leave Home (a non-event → a bounded wait).
 it('invalid id leaves the session undisturbed (stays Home)', async () => {
@@ -113,7 +80,7 @@ it('open (bare id) lands on the details springboard', async () => {
 // action=play → playback actually starts (the per-type quickplay engine).
 it('action=play starts playback', async () => {
   await castFromHome(`id=${heroId}|action=play`);
-  await waitMediaPlaying('play');
+  await waitMediaPlaying('deep-link play');
   await stopPlayback(); // leave a clean slate for the next spec
 }, 120000);
 
@@ -126,7 +93,7 @@ it('action=play starts playback', async () => {
 it('back-from-player does not re-fire; a repeat play cast does', async () => {
   // First play cast → the movie plays (Home → Details(?deeplink=play) → Player back-stack).
   await castFromHome(`id=${heroId}|action=play`);
-  await waitMediaPlaying('first play');
+  await waitMediaPlaying('deep-link first play');
 
   // Back from the player resumes the launching details (keepAlive). Same route.id, so
   // checkDeepLinkLaunch must NOT re-fire — playback must stay stopped.
@@ -144,7 +111,7 @@ it('back-from-player does not re-fire; a repeat play cast does', async () => {
   // Repeat cast of the SAME item, NO relaunch — we're on its loaded keepAlive details. A fresh
   // navigation route.id (the in-place playFromDeepLink path) must RE-fire playback.
   await ecp.sendInput({ params: { contentId: `id=${heroId}|action=play` } });
-  await waitMediaPlaying('repeat cast play');
+  await waitMediaPlaying('deep-link repeat cast play');
   await stopPlayback();
 }, 150000);
 
@@ -161,6 +128,6 @@ it('library/container id lands on its grid', async () => {
 // action=instantmix → builds an instant mix from the item and starts playing it.
 it('action=instantmix builds a mix and starts playback', async () => {
   await castFromHome(`id=${audioId}|action=instantmix`);
-  await waitMediaPlaying('instantmix');
+  await waitMediaPlaying('deep-link instantmix');
   await stopPlayback(); // audio keeps playing while browsing by design — stop it so it can't leak
 }, 120000);
